+++ test/test_pipelines.py
@@ -1,9 +1,10 @@
 import transformers
 
+import gradio as gr
+
 class TestLoadFromPipeline:
     def test_text_to_text_model_from_pipeline(self):
         pipe = transformers.pipeline(model="sshleifer/bart-tiny-random")
-        output = pipe("My name is Sylvain and I work at Hugging Face in Brooklyn")
-        self.assertIsNotNone(output)
-
-
+        io = gr.Interface.from_pipeline(pipe)
+        output = io("My name is Sylvain and I work at Hugging Face in Brooklyn")
+        assert isinstance(output, str)
